moved pkg/kafka to internal/kafka
