+++ cmd/ingester/app/consumer/consumer_test.go
@@ -21,8 +21,8 @@ import (
 	"github.com/jaegertracing/jaeger/cmd/ingester/app/processor"
 	pmocks "github.com/jaegertracing/jaeger/cmd/ingester/app/processor/mocks"
 	"github.com/jaegertracing/jaeger/internal/metricstest"
-	"github.com/jaegertracing/jaeger/pkg/kafka/consumer"
-	kmocks "github.com/jaegertracing/jaeger/pkg/kafka/consumer/mocks"
+	"github.com/jaegertracing/jaeger/internal/storage/kafka/consumer"
+	kmocks "github.com/jaegertracing/jaeger/internal/storage/kafka/consumer/mocks"
 	"github.com/jaegertracing/jaeger/pkg/metrics"
 )
