+++ cmd/ingester/app/consumer/processor_factory_test.go
@@ -14,7 +14,7 @@ import (
 
 	cmocks "github.com/jaegertracing/jaeger/cmd/ingester/app/consumer/mocks"
 	"github.com/jaegertracing/jaeger/cmd/ingester/app/processor/mocks"
-	kmocks "github.com/jaegertracing/jaeger/pkg/kafka/consumer/mocks"
+	kmocks "github.com/jaegertracing/jaeger/internal/storage/kafka/consumer/mocks"
 	"github.com/jaegertracing/jaeger/pkg/metrics"
 )
