+++ internal/storage/kafka/auth/options_test.go
