+++ internal/storage/kafka/auth/plaintext.go
