+++ internal/storage/kafka/auth/plaintext_test.go
