+++ internal/storage/kafka/consumer/config.go
@@ -11,7 +11,7 @@ import (
 	cluster "github.com/bsm/sarama-cluster"
 	"go.uber.org/zap"
 
-	"github.com/jaegertracing/jaeger/pkg/kafka/auth"
+	"github.com/jaegertracing/jaeger/internal/storage/kafka/auth"
 )
 
 // Consumer is an interface to features of Sarama that are necessary for the consumer
