+++ internal/storage/v1/kafka/factory_test.go
@@ -16,7 +16,7 @@ import (
 	"go.uber.org/zap/zapcore"
 
 	"github.com/jaegertracing/jaeger/internal/config"
-	kafkaConfig "github.com/jaegertracing/jaeger/pkg/kafka/producer"
+	kafkaConfig "github.com/jaegertracing/jaeger/internal/storage/kafka/producer"
 	"github.com/jaegertracing/jaeger/pkg/metrics"
 )
